python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-strategy-sdk"
version = "1.0.0"
authors = [
    {name = "Crypto Strategy Team", email = "team@cryptostrategy.dev"},
]
description = "Cryptocurrency Trading Strategy Development Kit for algorithmic trading"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["trading", "cryptocurrency", "strategy", "algorithmic", "bitcoin", "fintech", "quantitative"]

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "requests>=2.25.0",
    "PyYAML>=5.4.0",
    "python-dateutil>=2.8.0",
    "pytz>=2021.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
    "jupyter>=1.0.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
]
backtest = [
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "openpyxl>=3.0.0",
]
all = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
    "jupyter>=1.0.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "openpyxl>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/crypto-strategy/crypto-strategy-sdk"
Documentation = "https://crypto-strategy-sdk.readthedocs.io"
Repository = "https://github.com/crypto-strategy/crypto-strategy-sdk.git"
Issues = "https://github.com/crypto-strategy/crypto-strategy-sdk/issues"
Changelog = "https://github.com/crypto-strategy/crypto-strategy-sdk/blob/main/CHANGELOG.md"

[project.scripts]
crypto-strategy-cli = "crypto_strategy_sdk.cli:main"

[tool.setuptools]
packages = ["crypto_strategy_sdk"]

[tool.setuptools.package-data]
crypto_strategy_sdk = ["*.yaml", "*.json", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["crypto_strategy_sdk"]
omit = [
    "tests/*",
    "examples/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]